requirements.txt
# ── Fase 1: API REST + JWT ────────────────────────────────────
PyJWT[crypto]==2.9.0
alembic==1.13.2
Flask-Limiter==3.8.0
redis==5.0.8
pytest==8.3.3
pytest-flask==1.3.0

# ── Futuras fases (comentadas hasta que se necesiten) ─────────
# pywebview==5.4          # Fase 4: shell desktop
# pyinstaller==6.10       # Fase 4: empaque .exe
# pysqlcipher3==1.2       # Fase 4: SQLite cifrado
# apsw==3.45.1            # Fase 4: SQLite bajo nivel
# argon2-cffi==23.1.0     # Fase 4: KDF para SQLCipher
# keyring==24.3.1         # Fase 4: DPAPI / Keychain
# httpx[http2]==0.27.0    # Fase 4: HTTP/2 para sync worker

# ── Dependencias existentes ───────────────────────────────────
arabic-reshaper==3.0.0
asn1crypto==1.5.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
cssselect2==0.9.0
docopt==0.6.2
et_xmlfile==2.0.0
Flask==3.1.2
flask-cors==6.0.2
Flask-Mail==0.10.0
Flask-Reuploaded==1.4.0
Flask-WTF==1.2.2
freetype-py==2.5.1
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
html5lib==1.1
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
num2words==0.5.14
numpy==2.4.2
oauthlib==3.3.1
openpyxl==3.1.5
oscrypto==1.3.0
pandas==3.0.1
pillow==12.1.1
proto-plus==1.27.2
protobuf==7.34.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==3.0
pyHanko==0.33.0
pyhanko-certvalidator==0.29.1
pyparsing==3.3.2
pypdf==6.7.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
rlPyCairo==0.4.0
six==1.17.0
svglib==1.6.0
tinycss2==1.5.1
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
uritools==6.0.1
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.1.5
WTForms==3.2.1
xhtml2pdf==0.2.17
